Make WAF logging representation more compact by removing unneeded header.
