+++ apache2/apache2_config.c
@@ -29,7 +29,7 @@
 #include "msc_lua.h"
 #endif
 #ifdef WAF_JSON_LOGGING_ENABLE
-#include "waf_log_util_external.h"
+#include "waf_log_util.h"
 #endif
